Use @dolittle/tooling.common.host
