+++ package.json
@@ -49,6 +49,7 @@
     "@dolittle/tooling.common.utilities": "8.*"
   },
   "dependencies": {
+    "@dolittle/tooling.common.host": "8.0.0-alpha.9",
     "chalk": "2.4.2",
     "conf": "3.0.0",
     "debug": "4.1.1",
